backend/requirements.txt
# Flask Core
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
flask-mail==0.9.1
flask-limiter==3.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0

# Utilities
email-validator==2.1.0
bleach==6.1.0
python-dateutil==2.8.2
requests==2.31.0

# Production
gunicorn==21.2.0
flask-compress==1.14

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
